+++ LuckyoneUI/Locales/deDE.lua
@@ -179,6 +179,7 @@ L["Please click the button below to apply Luckyones profile for BigWigs and Litt
 L["Please click the button below to apply Luckyones profile for Details! Damage Meter."] = "Bitte klicke auf die Schaltfläche unten, um Luckyones Profile für Details! Damage Meter anzuwenden."
 L["Please click the button below to apply Luckyones profile for WarpDeplete."] = "Bitte klicke auf die Schaltfläche unten, um Luckyones Profile für WarpDeplete anzuwenden."
 L["Please read the steps carefully before clicking any buttons."] = "Bitte lies die Schritte sorgfältig durch, bevor du auf eine Schaltfläche klickst."
+L["Plumber profile has been set."] = "Plumber Profil wurde festgelegt."
 L["Previous"] = "Zurück"
 L["Priest"] = "Piester"
 L["Privacy Overlay"] = "Datenschutz-Overlay"