+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated game tooltip default font size (11 > 12)  
 
 **Retail:**  
+- Added Plumber profile to /lucky config  
 - Updated BuffReminders profile  
 - Updated PermoksAccountManager profile  
 - Use Blizzard auras by default for Nameplates (Disables custom filtering)